waste-system/backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# YOLO Detection
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.26.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
pillow==10.1.0
